bench: Make Xor benchmark more representative
To make the benchmarks representative, I've collected the write vector's sizes during IBD for every invocation of `util::Xor` until 860k blocks, and used it as a basis for the micro-benchmarks, having a similar distribution of random data.
The histogram is in a separate compressed file (storing the differences compared to the previous row, achieving 15x compression), which is converted during build to a decompressed histogram (i.e. how many utils::Xor calls were performed during IBD with vectors of the given size).

And even though we already have serialization tests, `AutoFileXor` was added to serializing 1 MB via the provided key_bytes. This was used to test the effect of disabling obfuscation.

>  cmake -B build -DBUILD_BENCH=ON -DCMAKE_BUILD_TYPE=Release \
&& cmake --build build -j$(nproc) \
&& build/src/bench/bench_bitcoin -filter='XorHistogram|AutoFileXor' -min-time=10000

C++ compiler .......................... AppleClang 16.0.0.16000026

|             ns/byte |              byte/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|                1.86 |      536,276,762.89 |    0.3% |     10.25 | `AutoFileXor`
|                1.04 |      961,221,881.72 |    0.2% |     93.30 | `XorHistogram`

C++ compiler .......................... GNU 13.2.0

|             ns/byte |              byte/s |    err% |        ins/byte |        cyc/byte |    IPC |       bra/byte |   miss% |     total | benchmark
|--------------------:|--------------------:|--------:|----------------:|----------------:|-------:|---------------:|--------:|----------:|:----------
|                1.83 |      547,257,936.65 |    0.2% |            9.20 |            3.47 |  2.655 |           1.03 |    0.1% |     11.04 | `AutoFileXor`
|                0.97 |    1,032,916,679.87 |    0.0% |            9.01 |            3.29 |  2.738 |           1.00 |    0.0% |     86.58 | `XorHistogram`