Set -DSECP256K1_EXPERIMENTAL=ON to fix ci build failure
SECP256K1_EXPERIMENTAL=ON is hardcoded in src/secp256k1/CMakeLists.txt:159 and I can build locally without the setting this option manually
But this doesn't work on the ci. I couldn't find the cause. I don't see any command overriding the default value
I added this option as a hack to fix the error