[refactor] validation: split out future timestamp check
In the following commits we'll make it so all consensus rules are checked when
connecting blocks. However we should not check again there that the block
timestamp is not more than 2 hours in the future, just in case our system clock
went backward.

To achieve this we first split out the future timestamp check in its own function.